+++ crates/platform/src/api/blocks/hash_or_number/txs.rs
@@ -1,9 +1,7 @@
+use crate::blocks::{BlockData, BlocksPath};
 use crate::{api::ApiResult, server::state::AppState};
 use axum::extract::{Path, Query, State};
-use bf_common::{
-    blocks::{BlockData, BlocksPath},
-    pagination::{Pagination, PaginationQuery},
-};
+use bf_common::pagination::{Pagination, PaginationQuery};
 
 pub async fn route(
     State(state): State<AppState>,