+++ crates/platform/src/api/txs/hash/root.rs
@@ -1,7 +1,7 @@
+use crate::txs::TxsPath;
 use crate::{api::ApiResult, server::state::AppState};
 use axum::extract::{Path, State};
 use bf_api_provider::types::TxsSingleResponse;
-use bf_common::txs::TxsPath;
 
 pub async fn route(
     State(state): State<AppState>,