+++ crates/platform/src/config.rs
@@ -1,8 +1,8 @@
 use crate::cli::Args;
-use crate::errors::AppError;
 use crate::genesis::{GenesisRegistry, genesis};
-use crate::types::Network;
 use bf_api_provider::types::GenesisResponse;
+use bf_common::errors::AppError;
+use bf_common::types::Network;
 use clap::ValueEnum;
 use futures::FutureExt; // for `.boxed()`
 use futures::future::BoxFuture;