+++ nix/internal/hydra-platform-gateway-test.sh
@@ -288,6 +288,7 @@ done
 # ---------------------------------------------------------------------------- #
 
 log_level=info
+peer_secret="hydra-platform-test-peer-secret"
 
 gateway_port=$(python3 -m portpicker)
 platform_port=$(python3 -m portpicker)
@@ -305,6 +306,7 @@ cat >gateway-config.toml <<EOF
 [server]
 address = '127.0.0.1:${gateway_port}'
 log_level = '${log_level}'
+peer_secret = '${peer_secret}'
 
 [database]
 connection_string = 'postgresql://not-used-with-dev-mock-db@localhost/dummy'